pyproject.toml
[tool.poetry]
name = "ha-social-network"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.15.2"
asyncpg = "^0.30.0"
bcrypt = "^4.3.0"
fastapi = "^0.115.12"
psycopg2-binary = "^2.9.10"
pydantic = {version = "^2.11.3", extras = ["email"]}
pydantic-settings = "^2.9.1"
PyJWT = "^2.10.1"
uvicorn = "^0.34.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"
httpx = "^0.28.1"
pytest = "7.4.4"
pytest-asyncio = "0.21.2"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
